requirements.txt
click==8.0.1
colorama==0.4.4
dataclasses==0.8
Flask==2.0.1
importlib-metadata==4.6.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pyFirmata==1.1.0
pyserial==3.5
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.4.1
